Error handling - timing out is not an error
The lambda timing out before all hotel ids have been processed
isn't really an error as long as some results have been pushed
back to the client. A lot of invocations *will* timeout because
the number of hotel ids to look at is large and the external API
is slow - which will then pollute the monitoring dashboard with
errors that aren't actionable.